% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_amelia.R
\name{patch_amelia}
\alias{patch_amelia}
\title{Impute gaps in daily weather data using multiple imputation of Amelia package}
\usage{
patch_amelia(
  weather,
  target,
  weather_info,
  rain_data = F,
  prcp_threshold = 1,
  n.impute = 5,
  parallel = NA
)
}
\arguments{
\item{weather}{data.frame with columns for each weather station and rows for
each daily observation. All columns need to contain observations of the same
variable. Missing observations need to be marked with NA}

\item{target}{character, column name in weather, on which the method should be applied}

\item{weather_info}{data.frame containing the name / id of the weather station (needs
to be the same as the column names in weather), Longitude and Latitude in
decimal format}

\item{rain_data}{logical, in case precipitation data should be patched the 
mice package allows to at first decide if precipitation occures and in a
second step to patch precipitation amount only in cases where the patching
algorithm expects precipitation to happen.}

\item{prcp_threshold}{numerical, setting the threshold until which
precipitation amount the algorithm treats observations are real precipitation
events. Values below are treated as zeros}

\item{parallel}{logical, if true the paralleled version Amelia::amelia is used}

\item{n_impute}{number of multiple imputations, default is 5}
}
\value{
same data.frame as weather but with all NAs imputed for all columns. There
can be still cases of NA, if for a certain observation none of the other stations
had valid observations
}
\description{
Takes data.frame of daily weather data of several neighboring weather stations
and imputes gaps with multiple imputation method of the Amelia package.
}
\details{
For more details of the function please refer to \code{\link{Amelia::amelia}}
}
\examples{
#think of example here
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
