% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_info.R
\docType{data}
\name{weather_info}
\alias{weather_info}
\title{Metadata of weather stations}
\format{
data.frame with 13 row and 4 variables:
\describe{
  \item{id}{unique identifies of weather station}
  \item{Name}{name of weather station}
  \item{Longitude}{longitude of weather station decimal format}
  \item{Latitude}{latitude of weather station in decimal format}
}
}
\source{
\url{https://cimis.water.ca.gov/} and \url{http://ipm.ucanr.edu/}
}
\usage{
data(weather_info)
}
\description{
data.frames containing a unique identifier for the weather station, coordinates
in decimal format and name of the weather startion.
}
\keyword{datasets}
